==> Part 1 Core.ipynb <==
# importing libraries
import pandas as pd
import numpy as np

## Filtering the data

# getting the data
basics_url = 'https://datasets.imdbws.com/title.basics.tsv.gz'
ratings_url = 'https://datasets.imdbws.com/title.ratings.tsv.gz'
akas_url = 'https://datasets.imdbws.com/title.akas.tsv.gz'

# converting the urls to dataframes
basics = pd.read_csv(basics_url, sep='\t', low_memory=False)

ratings = pd.read_csv(ratings_url, sep='\t', low_memory=False)

akas = pd.read_csv(akas_url, sep='\t', low_memory=False)

# replacing \N characters in the dataframe to numpy nan values
basics.replace({'\\N':np.nan}, inplace = True)
basics.info()

# looking at the data
basics.head()

# dropping rows that have a missing value in genres or runtimeMinutes
basics.dropna(subset=['genres', 'runtimeMinutes'], inplace=True)

# converting the startYear column to floats
basics['startYear'] = basics['startYear'].astype('float')

# filtering out to movies after 2000 and before 2022 inclusive
basics = basics.loc[(basics['startYear'] >= 2000) & (basics['startYear'] <= 2022) & (basics['titleType'] == 'movie')]

basics['startYear'] = basics['startYear'].astype('int32')
basics.info()

# filtering out documentaries
is_documentary = basics['genres'].str.contains('documentary',case=False)
basics = basics[~is_documentary]

# replacing \N characters in the dataframe to numpy nan values
akas.replace({'\\N':np.nan}, inplace = True)

# filtering to only include the US region
akas = akas.loc[akas['region'] == 'US']

# creating a filter for the basics dataframe using the akas dataframe to look
# for id's that appear in both. Since akas is already filtered to only be
# US this works to help filter basics the same way
keepers = basics['tconst'].isin(akas['titleId'])

basics = basics[keepers]

# replacing \N characters in the dataframe to numpy nan values
ratings.replace({'\\N':np.nan}, inplace = True)

# creating a filter for the ratings database using the akas database
keepers2 = ratings['tconst'].isin(akas['titleId'])

ratings = ratings[keepers2]

## File management and downloading

# Creating a file for the data
import os
os.makedirs('Data/',exist_ok=True) 
os.listdir("Data/")

# downloading the dataframes
basics.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)

basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)
basics.head()

ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)

ratings = pd.read_csv("Data/title_ratings.csv.gz", low_memory = False)
ratings.head()

akas.to_csv("Data/title_akas.csv.gz",compression='gzip',index=False)

akas = pd.read_csv("Data/title_akas.csv.gz", low_memory = False)
akas.head()

## Final dataframes

basics.info()

ratings.info()

akas.info()
