==> Part 2 Core.ipynb <==
# Importing libraries and creating data folder
import pandas as pd
import os, time, json
import tmdbsimple as tmdb
from tqdm.notebook import tqdm_notebook
FOLDER = 'Data/'
os.makedirs(FOLDER, exist_ok = True)
os.listdir(FOLDER)

# Accessing TMDB API
with open('/Users/Dido Admin/.secret/tmdb_api.json', 'r') as f:
    login = json.load(f)
tmdb.API_KEY = login['api-key']

# Method to grab the movie information as well as the rating of the movie.
def get_movie_with_rating(movie_id):
    # Get the movie object for the current id
    movie = tmdb.Movies(movie_id)
    # save the .info .releases dictionaries
    info = movie.info()
    releases = movie.releases()
    # Loop through countries in releases
    for c in releases['countries']:
        # if the country abbreviation==US
        if c['iso_3166_1' ] =='US':
            ## save a "certification" key in the info dict with the certification
            info['certification'] = c['certification']
    return info

# Appends a list of records to an already existing JSON file.
def write_json(new_data, filename): 
    """Appends a list of records (new_data) to a json file (filename). 
    Adapted from: https://www.geeksforgeeks.org/append-to-json-file-using-python/"""  
    
    with open(filename,'r+') as file:
        # First we load existing data into a dict.
        file_data = json.load(file)
        ## Choose extend or append
        if (type(new_data) == list) & (type(file_data) == list):
            file_data.extend(new_data)
        else:
             file_data.append(new_data)
        # Sets file's current position at offset.
        file.seek(0)
        # convert back to json.
        json.dump(file_data, file)

# Grabbing the data that was filtered in part 1.
basics = pd.read_csv('Data/title_basics.csv.gz')
basics.head()

# For this assignment the years 2000 and 2001 were asked for.
YEARS_TO_GET = [2000, 2001]
errors = []

# Loop to get the information one year at a time and store the data in
# seperate files for each year.
for YEAR in tqdm_notebook(YEARS_TO_GET, desc = 'YEARS', position = 0):
    # Create the path for the json file.
    JSON_FILE = f'{FOLDER}tmdb_api_results_{YEAR}.json'
    file_exists = os.path.isfile(JSON_FILE)
    
    # If it doesn't exist create it.
    if file_exists == False:
        with open(JSON_FILE, 'w') as f:
            json.dump([{'imdb_id': 0}], f)
    
    # Filter the basics dataframe to only include data from the current year we want.
    df = basics.loc[basics['startYear'] == YEAR].copy()
    # Grab the movie ids of those movies that meet the criteria.
    movie_ids = df['tconst'].copy()
    # Load previous data so we can skip over movies that we already have.
    previous_df = pd.read_json(JSON_FILE)
    # filter out ids that are already in the file.
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]
    
    for movie_id in tqdm_notebook(movie_ids_to_get,
                                  desc = f'Movies from {YEAR}',
                                  position = 1,
                                  leave = True):
        try:
            # Retrieve the data for the movie id
            temp = get_movie_with_rating(movie_id)
            write_json(temp, JSON_FILE)
            time.sleep(0.02)
        except Exception as e:
            errors.append([movie_id, e])
    final_year_df = pd.read_json(JSON_FILE)
    final_year_df.to_csv(f"{FOLDER}final_tmdb_data_{YEAR}.csv.gz", compression="gzip", index=False)

print(f"- Total errors: {len(errors)}")

# Testing the get_movie_with_rating method individually.
get_movie_with_rating("tt0848228")

get_movie_with_rating("tt0332280")
