==> Part 3 Core.ipynb <==
# Importing libraries
import pandas as pd
import numpy as np
import json

## Transforming Data

# Importing data
basics_df = pd.read_csv('Data/title_basics.csv.gz')
basics_df.head()

ratings_df = pd.read_csv('Data/title_ratings.csv.gz')
ratings_df.head()

# Splitting the genres column
basics_df['genres_split'] = basics_df['genres'].str.split(',')
basics_df

# Explodiung the split column
exploded_genres = basics_df.explode('genres_split')
exploded_genres

# Saving the unique values for a table in the future.
unique_genres = sorted(exploded_genres['genres_split'].unique())

# Creating a title_genres dataframe
title_genres = exploded_genres[['tconst', 'genres_split']]
title_genres.head()

# Making the genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

# make new integer genre_id and drop string genres
title_genres['genre_id'] = title_genres['genres_split'].map(genre_map)
title_genres = title_genres.drop(columns='genres_split')

title_genres.head()

# Creating the genres dataframe
genres = pd.DataFrame({'Genre_Name': genre_map.keys(),
                      'Genre_ID':  genre_map.values()})
genres.head()

basics_df.head()

# Dropping unnecessary columns.
basics_df.drop(columns =  ['originalTitle', 'isAdult', 'titleType', 'genres', 'genres_split'], inplace = True)
basics_df.head()

## MySQL Database Creation

# Importing libraries
import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy_utils import create_database, database_exists
from sqlalchemy.types import *
import json

# Accessing my mysql information
with open('/Users/Dido Admin/.secret/mysql.json') as f:
    login = json.load(f)
login.keys()

# Creating the mysql connection
connection_str = f"mysql+pymysql://{login['username']}:{login['password']}@localhost/movies"

engine = create_engine(connection_str)

# Creating the database
if database_exists(connection_str) == False:
  create_database(connection_str)
else:
  print('The database already exists')

database_exists(connection_str)

# Calculate max string lengths for object columns
key_len = basics_df['tconst'].fillna('').map(len).max()
title_len = basics_df['primaryTitle'].fillna('').map(len).max()
# Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    "primaryTitle": Text(title_len+1),
    'startYear':Float(),
    'endYear':Float(),
    'runtimeMinutes':Integer()}

# Turning the dataframe into a sql table
basics_df.to_sql('title_basics', engine, dtype = df_schema, if_exists = 'replace', index=False)

# Adding a primary key
engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')

# Displaying the table
q = """SELECT* FROM title_basics LIMIT 5;"""
pd.read_sql(q, engine)

Most of the comments of the later code cells would be very similar. I will only comment if something new happens.

df_schema = {
    "tconst": String(key_len+1), 
    'averageRating':Float(),
    'numVotes':Integer()}

ratings_df.to_sql('title_ratings', engine, dtype = df_schema, if_exists = 'replace', index=False)

engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')

q = """SELECT* FROM title_ratings LIMIT 5;"""
pd.read_sql(q, engine)

df_schema = {
    "tconst": String(key_len+1), 
    'genre_id':Integer()}

title_genres.to_sql('title_genres', engine, dtype = df_schema, if_exists = 'replace', index=False)

q = """SELECT* FROM title_genres LIMIT 5;"""
pd.read_sql(q, engine)

name_len = genres['Genre_Name'].fillna('').map(len).max()
df_schema = {
    "Genre_Name": String(name_len+1), 
    'Genre_ID':Integer()}

genres.to_sql('genres', engine, dtype = df_schema, if_exists = 'replace', index=False)

engine.execute('ALTER TABLE genres ADD PRIMARY KEY (`Genre_ID`);')

q = """SELECT* FROM genres LIMIT 5;"""
pd.read_sql(q, engine)

# Importing previous data check out Part 2 for more information.
tmdb_data_df = pd.read_csv('Data/tmdb_results_combined.csv.gz')
tmdb_data_df.head()

# Removing first row
tmdb_data_df = tmdb_data_df.loc[1:,]

tmdb_data_df.info()

# This is the schema dictionary again.
imdb_id_len = tmdb_data_df['imdb_id'].fillna('').map(len).max()
backdrop_path_len = tmdb_data_df['backdrop_path'].fillna('').map(len).max()
belongs_to_collection_len = tmdb_data_df['belongs_to_collection'].fillna('').map(len).max()
genres_len = tmdb_data_df['genres'].fillna('').map(len).max()
homepage_len = tmdb_data_df['homepage'].fillna('').map(len).max()
original_language_len = tmdb_data_df['original_language'].fillna('').map(len).max()
original_title_len = tmdb_data_df['original_title'].fillna('').map(len).max()
overview_len = tmdb_data_df['overview'].fillna('').map(len).max()
poster_path_len = tmdb_data_df['poster_path'].fillna('').map(len).max()
production_companies_len = tmdb_data_df['production_companies'].fillna('').map(len).max()
production_countries_len = tmdb_data_df['production_countries'].fillna('').map(len).max()
release_date_len = tmdb_data_df['release_date'].fillna('').map(len).max()
spoken_languages_len = tmdb_data_df['spoken_languages'].fillna('').map(len).max()
status_len = tmdb_data_df['status'].fillna('').map(len).max()
tagline_len = tmdb_data_df['tagline'].fillna('').map(len).max()
title_len = tmdb_data_df['title'].fillna('').map(len).max()
certification_len = tmdb_data_df['certification'].fillna('').map(len).max()

df_schema = {
    "imdb_id": String(imdb_id_len+1), 
    "adult":Float(),
    "backdrop_path": String(backdrop_path_len+1), 
    "belongs_to_collection": String(belongs_to_collection_len+1), 
    "budget":Float(),
    "genres": String(genres_len+1),
    "homepage": String(homepage_len+1),
    "id":Float(),
    "original_language": String(original_language_len+1),
    "original_title": String(original_title_len+1),
    "overview": String(overview_len+1),
    "popularity":Float(),
    "poster_path": String(poster_path_len+1),
    "production_companies": String(production_companies_len+1),
    "production_countries": String(production_countries_len+1),
    "release_date": String(release_date_len+1),
    "revenue":Float(),
    "runtime":Float(),
    "spoken_languages": String(spoken_languages_len+1),
    "status": String(status_len+1),
    "tagline": String(tagline_len+1),
    "title": String(title_len+1),
    "video":Float(),
    "vote_average":Float(),
    "vote_count":Float(),
    "certification": String(certification_len+1)}

tmdb_data_df.to_sql('tmdb_data', engine, dtype = df_schema, if_exists = 'replace', index=False)

engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

q = """SELECT* FROM tmdb_data LIMIT 5;"""
pd.read_sql(q, engine)

#  Showing all the tables
q = """SHOW TABLES;"""
pd.read_sql(q, engine)
